requirements.txt
aiogram~=2.25.1
python-dotenv~=1.0.0
openai~=0.27.8
psycopg2~=2.9.6